Removes custom read syntax in favour of procedure
This is done because the test runner cries if you use the custom read
syntax; i.e. calling #u8(10 20 30) instead of (u8vector 10 20 30). This
is probably something to do with how CHICKEN sets custom read syntax for
the REPL (csi) but not for code evaluated during a test. In any case,
it's not worth trying to hang on to the custom read syntax when I can
just swap it out for the procedure call to fix the error.